Fixed and refactored MyBatis configurations
